requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
transformers==4.44.2
torch==2.7.0
pillow==10.4.0
python-multipart